Added solution - LeetHub

+++ Minimum Cost of ropes - GFG/minimum-cost-of-ropes.cpp
@@ -15,7 +15,7 @@ class Solution
         //push all the elements into min-heap
         priority_queue<long long,vector<long long>,greater<long long>> pq;
         
-        for(long long i = 0;i<n;i++)
+        for(int i = 0;i<n;i++)
             pq.push(arr[i]);
